+++ lib/src/account.dart
@@ -1,16 +1,20 @@
+import 'package:equatable/equatable.dart';
+import 'package:meta/meta.dart';
 import 'request_builder.dart';
 import 'auth.dart';
 
-abstract class Account {
-  String type;
-  String id;
+@immutable
+abstract class Account extends Equatable {
+  const Account({@required this.type, @required this.id});
+  final String type;
+  final String id;
 
   RequestBuilder sign(RequestBuilder requestBuilder);
 
-  static void fromJson(Account account, Map<dynamic, dynamic> json) {
-    account.type = json["type"]?.toString() ?? "";
-    account.id = json["id"]?.toString() ?? "";
-  }
+  // static void fromJson(Account account, Map<dynamic, dynamic> json) {
+  //   account.type = json["type"]?.toString() ?? "";
+  //   account.id = json["id"]?.toString() ?? "";
+  // }
 
   Map<dynamic, dynamic> toJson() {
     final output = Map<dynamic, dynamic>();
@@ -18,10 +22,17 @@ abstract class Account {
     output["id"] = id;
     return output;
   }
+
+  @override
+  List<Object> get props => [type, id];
 }
 
+@immutable
 abstract class OAuth2Account extends Account {
-  String accessToken;
+  const OAuth2Account({this.accessToken, String type, String id})
+      : super(type: type, id: id);
+
+  final String accessToken;
 
   String authorizeUrl();
   String callbackUrl();
@@ -32,14 +43,15 @@ abstract class OAuth2Account extends Account {
     throw UnimplementedError("OAuth2 flow not implemented");
   }
 
-  void setAuth(Auth auth) {
-    accessToken = auth.access_token;
-  }
+  OAuth2Account copyWithAuth(Auth auth);
+  // {
+  //   accessToken = auth.access_token;
+  // }
 
-  static void fromJson(OAuth2Account account, Map<dynamic, dynamic> json) {
-    Account.fromJson(account, json);
-    account.accessToken = json["accessToken"]?.toString() ?? "";
-  }
+  // static void fromJson(OAuth2Account account, Map<dynamic, dynamic> json) {
+  //   Account.fromJson(account, json);
+  //   account.accessToken = json["accessToken"]?.toString() ?? "";
+  // }
 
   @override
   Map<dynamic, dynamic> toJson() {
@@ -47,4 +59,7 @@ abstract class OAuth2Account extends Account {
     output["accessToken"] = accessToken;
     return output;
   }
+
+  @override
+  List<Object> get props => [...super.props, accessToken];
 }